+++ pennylane_lightning/core/src/utils/Util.hpp
@@ -574,4 +574,18 @@ std::vector<T1> cast_vector(const std::vector<T0> &vec) {
     return result;
 }
 
+/**
+ * @brief Check if two vectors are disjoint.
+ * @tparam T Data type.
+ * @param v1 First vector.
+ * @param v2 Second vector.
+ *
+ * @return bool True if the vectors are disjoint, false otherwise.
+ */
+template <typename T = std::size_t>
+bool areVecsDisjoint(const std::vector<T> &v1, const std::vector<T> &v2) {
+    std::set<std::size_t> s0(v1.begin(), v1.end());
+    s0.insert(v2.begin(), v2.end());
+    return s0.size() == v1.size() + v2.size();
+}
 } // namespace Pennylane::Util